+++ Example/Source/Controllers/OptionsController/OptionsView.swift
@@ -11,7 +11,7 @@ import UIKit
 protocol OptionsViewProtocol: class {
     var horizontalDemoButton: Actionable { get }
     var verticalDemoButton: Actionable { get }
-    var circularDemoButton: Actionable { get }
+    var carouselDemoButton: Actionable { get }
 }
 
 class OptionsView: UIView {
@@ -19,11 +19,12 @@ class OptionsView: UIView {
     private let optionButtonHeigh: CGFloat = 32
     private let horizontalDemoButtonCenterYOffset: CGFloat = -32
     private let verticalDemoButtonCenterYOffset: CGFloat = 0
-    private let circularDemoButtonCenterYOffset: CGFloat = 32
+    private let carouselDemoButtonCenterYOffset: CGFloat = 32
     private let internalHorizontalDemoButton = FilledButton()
     private let internalVerticalDemoButton = FilledButton()
-    private let internalCircularDemoButton = FilledButton()
-    private let copyrightLabel = UILabel()
+    private let internalCarouselDemoButton = FilledButton()
+    private let logoImageView = UIImageView()
+    private let label = UILabel()
     
     init() {
         super.init(frame: CGRect.zero)
@@ -43,17 +44,24 @@ class OptionsView: UIView {
         addSubview(internalVerticalDemoButton)
         activateOptionButtonConstraints(view: internalVerticalDemoButton, centerYOffset: verticalDemoButtonCenterYOffset)
         
-        internalCircularDemoButton.setTitle(NSLocalizedString("CircularSampleButtonTitle", comment: ""), for: .normal)
-        internalCircularDemoButton.clipsToBounds = true
-        internalCircularDemoButton.layer.cornerRadius = optionButtonHeigh / 2
-        internalCircularDemoButton.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(internalCircularDemoButton)
-        activateOptionButtonConstraints(view: internalCircularDemoButton, centerYOffset: circularDemoButtonCenterYOffset)
+        internalCarouselDemoButton.setTitle(NSLocalizedString("CarouselSampleButtonTitle", comment: ""), for: .normal)
+        internalCarouselDemoButton.clipsToBounds = true
+        internalCarouselDemoButton.layer.cornerRadius = optionButtonHeigh / 2
+        internalCarouselDemoButton.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(internalCarouselDemoButton)
+        activateOptionButtonConstraints(view: internalCarouselDemoButton, centerYOffset: carouselDemoButtonCenterYOffset)
 
-        copyrightLabel.text = NSLocalizedString("CopyrightText", comment: "")
-        copyrightLabel.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(copyrightLabel)
-        activateCopyrightLabelConstraints(view: copyrightLabel)
+        logoImageView.image = UIImage(named: "main_logo")
+        logoImageView.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(logoImageView)
+        activateLogoImageConstraints(view: logoImageView, anchorView: internalCarouselDemoButton)
+        
+        label.text = "SlideController"
+        label.font = UIFont.boldSystemFont(ofSize: 24)
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.textColor = UIColor(red: 61 / 255, green: 86 / 255, blue: 166 / 255, alpha: 1)
+        addSubview(label)
+        activateLabelConstraints(view: label)
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -75,13 +83,24 @@ private extension PrivateOptionsView {
             ])
     }
     
-    func activateCopyrightLabelConstraints(view: UIView) {
+    func activateLogoImageConstraints(view: UIView, anchorView: UIView) {
         guard let superview = view.superview else {
             return
         }
         NSLayoutConstraint.activate([
             view.centerXAnchor.constraint(equalTo: superview.centerXAnchor),
-            view.bottomAnchor.constraint(equalTo: superview.bottomAnchor, constant: -20)
+            view.bottomAnchor.constraint(equalTo: superview.bottomAnchor, constant: -20),
+            view.heightAnchor.constraint(equalToConstant: 60)
+            ])
+    }
+    
+    func activateLabelConstraints(view: UIView) {
+        guard let superview = view.superview else {
+            return
+        }
+        NSLayoutConstraint.activate([
+            view.centerXAnchor.constraint(equalTo: superview.centerXAnchor),
+            view.topAnchor.constraint(equalTo: superview.topAnchor, constant: 20)
             ])
     }
 }
@@ -96,7 +115,7 @@ extension OptionsViewProtocolImplementation: OptionsViewProtocol {
         return internalVerticalDemoButton
     }
     
-    var circularDemoButton: Actionable {
-        return internalCircularDemoButton
+    var carouselDemoButton: Actionable {
+        return internalCarouselDemoButton
     }
 }