+++ Example/Source/Protocols/ContentActionable.swift
@@ -13,7 +13,6 @@ protocol ContentActionable {
     var removeDidTapAction: Action? { get set }
     var insertDidTapAction: Action? { get set }
     var appendDidTapAction: Action? { get set }
-    var menuDidTapAction: Action? { get set }
     var changePositionAction: ((Int) -> ())? { get set }
     var changeTitleModeAction: ((Int) -> ())? { get set }
 }